backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via edge-tts
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
bidict==0.23.1
    # via python-socketio
certifi==2026.1.4
    # via
    #   edge-tts
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
edge-tts==7.2.7
    # via vibesync-backend (pyproject.toml)
fastapi==0.128.0
    # via vibesync-backend (pyproject.toml)
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via vibesync-backend (pyproject.toml)
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via
    #   vibesync-backend (pyproject.toml)
    #   uvicorn
python-engineio==4.13.0
    # via python-socketio
python-socketio==5.16.0
    # via vibesync-backend (pyproject.toml)
pyyaml==6.0.3
    # via uvicorn
redis==7.1.0
    # via spotipy
requests==2.32.5
    # via spotipy
simple-websocket==1.1.0
    # via python-engineio
spotipy==2.25.2
    # via vibesync-backend (pyproject.toml)
starlette==0.50.0
    # via fastapi
tabulate==0.9.0
    # via edge-tts
typing-extensions==4.15.0
    # via
    #   edge-tts
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via
    #   requests
    #   spotipy
uvicorn==0.40.0
    # via vibesync-backend (pyproject.toml)
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
wsproto==1.3.2
    # via simple-websocket
yarl==1.22.0
    # via aiohttp
